Move extract_extents into KokkosFFT_Helpers.hpp
