+++ common/src/KokkosFFT_Helpers.hpp
@@ -7,6 +7,7 @@
 
 #include <Kokkos_Core.hpp>
 #include "KokkosFFT_Common_Types.hpp"
+#include "KokkosFFT_Extents.hpp"
 #include "KokkosFFT_Layout.hpp"
 #include "KokkosFFT_MDOperations.hpp"
 #include "KokkosFFT_Traits.hpp"
